changes baseline_search to use separate module function to perform search for patching api request while testing. Using snake_casing. Add fixture files to tests/fixtures
